requirements.txt
python-dotenv==1.0.0
concurrent_log_handler==0.9.25
cryptography==41.0.7
Django==5.0.1
django_hosts==6.0
django_user_agents==0.4.0
django-extensions==3.2.3
djangorestframework==3.14.0
django-sslserver==0.22
djangorestframework_simplejwt==5.3.1
# rest_framework_simplejwt==0.0.2
django-cors-headers==4.3.0
django-safedelete==1.3.3
dj-database-url==2.1.0
psycopg2-binary==2.9.6 
django-phonenumber-field==7.3.0
django-redis==5.4.0
whitenoise==6.5.0
phonenumbers==8.13.22
docx==0.2.4
geoip2==4.8.0
Pillow==10.2.0
PyJWT==2.8.0
pytz==2023.3.post1
Requests==2.31.0
safedelete==0.0.2
SocksiPy_branch==1.1
SocksiPy_branch==1.01
twilio==8.11.0
user_agents==2.2.0
